Move DBALEventStoreException to CultuurNet\UDB3\EventSourcing\DBAL namespace
